% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{norm_DESeq2}
\alias{norm_DESeq2}
\title{DESeq2 normalisation}
\usage{
norm_DESeq2(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
DESeq2 normalisation
}
