% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linnorm.R
\name{norm_linnorm}
\alias{norm_linnorm}
\title{Linnorm normalisation}
\usage{
norm_linnorm(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
Linnorm normalisation
}
