% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcpm.R
\name{norm_logCPM}
\alias{norm_logCPM}
\title{DESeq2 normalisation}
\usage{
norm_logCPM(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
DESeq2 normalisation
}
