% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scnorm.R
\name{norm_scnorm}
\alias{norm_scnorm}
\title{SCnorm normalisation}
\usage{
norm_scnorm(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
SCnorm normalisation
}
