% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scran.R
\name{norm_scran}
\alias{norm_scran}
\title{Scran normalisation}
\usage{
norm_scran(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
Scran normalisation
}
