% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm.R, R/tmmwzp.R
\name{norm_TMM}
\alias{norm_TMM}
\title{TMM normalisation}
\usage{
norm_TMM(sce)

norm_TMM(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
TMM normalisation

TMM with zero pairing normalisation
}
