% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmmwzp.R
\name{norm_tmmwzp}
\alias{norm_tmmwzp}
\title{TMM with zero pairing normalisation}
\usage{
norm_tmmwzp(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
TMM with zero pairing normalisation
}
