% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_size.R
\name{normalise_lib_size}
\alias{normalise_lib_size}
\alias{normalize_lib_size}
\title{Library size normalise and transform to log scale}
\usage{
normalise_lib_size(x, log = TRUE, offset = 1)

normalize_lib_size(x, log = TRUE, offset = 1)
}
\arguments{
\item{x}{the SingleCellExperiment or count matrix}

\item{log}{whether output should be log-scale}

\item{offset}{the counts to add before taking log}
}
\value{
matrix of normalised values
}
\description{
Library size normalise and transform to log scale
}
