% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scran.R
\name{normalise_scran}
\alias{normalise_scran}
\alias{normalize_scran}
\title{Scran normalisation}
\usage{
normalise_scran(x, log = TRUE)

normalize_scran(x, log = TRUE)
}
\arguments{
\item{x}{the SingleCellExperiment object}

\item{log}{TRUE if normalised values should be returned on log-scale}
}
\value{

}
\description{
Scran normalisation
}
