% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcpm.R
\name{norm_logcpm}
\alias{norm_logcpm}
\title{Log-cpm normalisation}
\usage{
norm_logcpm(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
Log-cpm normalisation
}
