% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm.R
\name{norm_tmm}
\alias{norm_tmm}
\title{TMM normalisation}
\usage{
norm_tmm(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object with logcounts slot containing normalised expression values
}
\description{
TMM normalisation
}
